WebServer: add support of Authorization: Bearer in HTTP header
This patch implement some parts of RFC 6750 OAuth 2.0 Bearer Token Usage.
Authorization of resource access through the use of a token is limited to
token presence in header fields of HTTP request.

Even if expiration date is not mandatory within tokens, i think most of the time
tokens embeds such data to grant access to resources for a discrete period.
That's why this patch focus on searching for an expiration date within the token,
and store it in a map in case of successful token verification, to speed up next
requests processing with same token.

Since content of tokens is not standardized, callbacks must be provided to keep
the token decode and checks outside of libnavajo.

Signed-off-by: Julien Courtat <julien.courtat@aqsacom.com>